SCF049 and 055 consensus sequence + change file in which all the consensus sequence of all the markers that I sequenced are stored
psbA for 049 and 055 has been aligned among Rev and Fwd sequences. The consensus sequence created for 049, for 055 is just the reverse sequence. Trimming in the conserved region. Creation of a new fasta file called final_sequences.fasta in which all the sequences per marker per species that I've generated are store. This is the input file for the alignment/concatenation code.